Update types and refactor resource management
Switch GPUStencilValue to UInt for correctness. Replace 'confined' with 'memoryScope' for better resource handling. Refactor memory mappings and callbacks in Buffer and RenderPassDescriptor for improved efficiency and readability.

+++ wgpu4k/src/commonMain/kotlin/io.ygdrasil.wgpu/Types.kt
@@ -36,7 +36,7 @@ typealias GPUSize64 = ULong
 
 typealias GPUSize64Out = Long
 
-typealias GPUStencilValue = Long
+typealias GPUStencilValue = UInt
 
 typealias GPUTextureUsageFlags = Int
